Refactor XORTrialAccessContext into OneVarTrialAccessContext
