+++ harness/tests/experiment/fixtures/pytorch_onevar_model.py
@@ -42,6 +42,7 @@ import yaml
 
 from determined import experimental, pytorch
 from determined.pytorch import samplers
+from tests.experiment.fixtures.pytorch_xor_model import StepableLRSchedule
 
 try:
     import apex
@@ -96,16 +97,18 @@ def triangle_label_sum(updates: List) -> Any:
     return out
 
 
+def get_onevar_model() -> torch.nn.Module:
+    model = torch.nn.Linear(1, 1, False)
+    # Manually initialize the one weight to 0.
+    model.weight.data.fill_(0)
+    return model
+
+
 class BaseOneVarTrial(pytorch.PyTorchTrial):
     def __init__(self, context: pytorch.PyTorchTrialContext) -> None:
         self.context = context
 
-        model = torch.nn.Linear(1, 1, False)
-
-        # Manually initialize the one weight to 0.
-        model.weight.data.fill_(0)
-
-        self.model = context.wrap_model(model)
+        self.model = context.wrap_model(get_onevar_model())
 
         self.lr = 0.001
 
@@ -421,6 +424,61 @@ class OneVarTrialCustomEval(BaseOneVarTrial):
             loss_sum += self.loss_fn(output, labels)
 
         loss = loss_sum / len(data_loader)
+        return {"val_loss": loss}
+
+
+class OneVarTrialAccessContext(BaseOneVarTrial):
+    _searcher_metric = "loss"
+
+    def __init__(self, context: pytorch.PyTorchTrialContext) -> None:
+        super().__init__(context)
+
+        self.model_a = self.context.wrap_model(get_onevar_model())
+        self.model_b = self.context.wrap_model(get_onevar_model())
+        self.opt_a = self.context.wrap_optimizer(
+            torch.optim.SGD(self.model_a.parameters(), self.context.get_hparam("learning_rate"))
+        )
+        self.opt_b = self.context.wrap_optimizer(
+            torch.optim.SGD(self.model_b.parameters(), self.context.get_hparam("learning_rate"))
+        )
+        self.lrs_a = self.context.wrap_lr_scheduler(
+            StepableLRSchedule(self.opt_a),
+            step_mode=pytorch.LRScheduler.StepMode(
+                self.context.get_hparam("lr_scheduler_step_mode")
+            ),
+        )
+        self.lrs_b = self.context.wrap_lr_scheduler(
+            StepableLRSchedule(self.opt_b),
+            step_mode=pytorch.LRScheduler.StepMode(
+                self.context.get_hparam("lr_scheduler_step_mode")
+            ),
+        )
+
+    def train_batch(
+        self, batch: pytorch.TorchData, epoch_idx: int, batch_idx: int
+    ) -> Dict[str, torch.Tensor]:
+        assert self.context.models
+        assert self.context.optimizers
+        assert self.context.lr_schedulers
+
+        data, labels = batch
+        output = self.model_a(data)
+        loss = torch.nn.functional.binary_cross_entropy(output, labels.contiguous().view(-1, 1))
+
+        self.context.backward(loss)
+        self.context.step_optimizer(self.opt_a)
+
+        return {"loss": loss}
+
+    def evaluate_batch(self, batch: pytorch.TorchData, batch_idx: int) -> Dict[str, Any]:
+        assert self.context.models
+        assert self.context.optimizers
+        assert self.context.lr_schedulers
+
+        data, labels = batch
+        output = self.model_a(data)
+        loss = self.loss_fn(output, labels)
+
         return {"val_loss": loss}
